feat: add session edit route
