+++ src/components/atoms/types.ts
@@ -1,6 +1,6 @@
 export enum RowType {
   Select = "select",
   Detail = "detail",
-  Switch = "switch",
+  Toggle = "toggle",
   Default = "default",
 }
